Break distributed config import cycle
